requirements.txt
langchain==1.2.4
langgraph==1.0.6
psycopg2==2.9.11
asyncpg==0.30.0
fastapi==0.128.0

uvicorn==0.40.0